+++ .gitignore
@@ -1,7 +1,8 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-package-lock.json
+node_modules/
+package-lock.json/
 /.pnp
 .pnp.js
