+++ client/src/components/NavBar.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function NavBar() {
+
+
+    return (
+        <>
+        </>
+    )
+}
