--- client/src/components/nav-bar/NavBar.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import s from './NavBar.jsx'
-import logo from '../../img/logo.png';
-
-export default function NavBar() {
-
-
-    return 
-}
