+++ client/src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import NavBar from '../../components/nav-bar/NavBar';
-import Landing from '../../components/landing/Landing';
+import NavBar from '../components/NavBar';
+import Landing from '../components/Landing';
 
 export default function Home(){
     return (
